v_3071.1647

+++ NativePatcher/BridgeBuilder/Patcher_ExtCode/myext/mycef.cc
@@ -12,31 +12,31 @@ void MethodArgs::SetArgAsString(int argIndex, const wchar_t* str)
 	case 0:
 	{
 		this->arg0.type = JSVALUE_TYPE_STRING;
-		this->arg0.length = wcslen(str);
+		this->arg0.length = (int32_t)wcslen(str);
 		this->arg0.value.str = (uint16_t*)str;
 	}break;
 	case 1:
 	{
 		this->arg1.type = JSVALUE_TYPE_STRING;
-		this->arg1.length = wcslen(str);
+		this->arg1.length = (int32_t) wcslen(str);
 		this->arg1.value.str = (uint16_t*)str;
 	}break;
 	case 2:
 	{
 		this->arg2.type = JSVALUE_TYPE_STRING;
-		this->arg2.length = wcslen(str);
+		this->arg2.length = (int32_t)wcslen(str);
 		this->arg2.value.str = (uint16_t*)str;
 	}break;
 	case 3:
 	{
 		this->arg3.type = JSVALUE_TYPE_STRING;
-		this->arg3.length = wcslen(str);
+		this->arg3.length = (int32_t)wcslen(str);
 		this->arg3.value.str = (uint16_t*)str;
 	}break;
 	case 4:
 	{
 		this->arg4.type = JSVALUE_TYPE_STRING;
-		this->arg4.length = wcslen(str);
+		this->arg4.length = (int32_t)wcslen(str);
 		this->arg4.value.str = (uint16_t*)str;
 	}break;
 	}
